refactored editor.js more, creating two new files, moved grid related stuff to editorGridHandler.js and json import export stuff to editorJsonHandler.js.
